move original nodegroups apis to 040-node-manager/hooks/internal
