+++ modules/040-node-manager/hooks/internal/v1/boilerplate.go.txt
