+++ modules/040-node-manager/hooks/internal/v1/doc.go
