+++ modules/040-node-manager/hooks/internal/v1/nodegroup_generated.deepcopy.go
