+++ modules/040-node-manager/hooks/internal/v1alpha2/nodegroup.go
